// src/payload-types.ts
export interface Media {
  id: number;
  alt?: string | null;
  caption?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "bulletins".
 */
export interface Bulletin {
  id: number;
  title: string;
  category: 'weekly' | 'condolence' | 'announcement' | 'feast';
  publishDate: string;
  content?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  pdfAttachment?: (number | null) | Media;
  isUrgent?: boolean | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "events".
 */
export interface Event {
  id: number;
  title: string;
  startTime: string;
  endTime: string;
  eventType: 'liturgy' | 'vespers' | 'youth' | 'bible_study' | 'special';
  description?: string | null;
  location?: string | null;
  updatedAt: string;
  createdAt: string;
}